feat/color-helper: made color helpers
