+++ src/components/vuestic-utilities/color-helpers.demo.vue
@@ -1,0 +1,21 @@
+<template>
+  <VbDemo>
+    <VbCard>
+      <h3>Background utilities</h3>
+      <p class="background--danger">
+        .background--danger
+      </p>
+
+      <h3>Text color utilities</h3>
+      <p class="text--danger">
+        .text--danger
+      </p>
+    </VbCard>
+  </VbDemo>
+</template>
+
+<script>
+import './color-helpers-plugin'
+export default {
+}
+</script>
