--- src/components/vuestic-utilities/utilities.demo.vue
@@ -1,20 +1,0 @@
-<template>
-  <div>
-    <h3>Background utilities</h3>
-    <p class="background--danger">
-      .background--danger
-    </p>
-
-    <h3>Text color utilities</h3>
-    <p class="text--danger">
-      .text--danger
-    </p>
-  </div>
-</template>
-
-<script>
-import './index'
-export default {
-
-}
-</script>
