+++ src/vue-book/book-main.js
@@ -9,6 +9,7 @@ import { BusPlugin } from 'vue-epic-bus'
 import { registerVuesticObject } from '../components/resize-events'
 import { DropdownPopperPlugin } from '../components/vuestic-components/va-dropdown/dropdown-popover-subplugin'
 import { installPlatform } from '../components/vuestic-components/va-popup/install'
+import ColorHelpersPlugin from '../components/vuestic-utilities/color-helpers-plugin'
 
 // eslint-disable-next-line
 console.log(`Version: ${VERSION}, ${TIMESTAMP}, commit: ${COMMIT}`)
@@ -41,6 +42,7 @@ registerVuesticObject(Vue)
 Vue.use(BusPlugin)
 Vue.use(VueClipboard)
 Vue.use(DropdownPopperPlugin)
+Vue.use(ColorHelpersPlugin)
 
 new Vue({
   router,
